requirements.txt
requests
requests-mock
PyQt5
python-dotenv
cryptography